+++ lang/de/permissions.php
@@ -1,0 +1,13 @@
+<?php
+return [
+    'Clear Selection' => 'Alle abwählen',
+    'Guard Name' => 'Guard Name',
+    'Id' => 'ID',
+    'Name' => 'Name',
+    'Permission' => 'Berechtigung',
+    'Permissions' => 'Berechtigungen',
+    'Role' => 'Rolle',
+    'Roles' => 'Rollen',
+    'Select all' => 'Alle auswählen',
+    'Users' => 'Benutzer'
+];