+++ lang/en/permissions.php
@@ -1,0 +1,13 @@
+<?php
+return [
+    'Clear Selection' => 'Clear Selection',
+    'Guard Name' => 'Guard Name',
+    'Id' => 'Id',
+    'Name' => 'Name',
+    'Permission' => 'Permission',
+    'Permissions' => 'Permissions',
+    'Role' => 'Role',
+    'Roles' => 'Roles',
+    'Select all' => 'Select all',
+    'Users' => 'Users'
+];