+++ resources/js/tool.js
@@ -1,9 +1,12 @@
-Nova.booting((Vue, router) => {
-    router.addRoutes([
-        {
-            name: ':package_name',
-            path: '/:package_name',
-            component: require('./components/Tool').default,
-        },
-    ])
+import Tool from './components/Tool'
+import IndexField from "./components/IndexField.vue";
+import DetailField from "./components/DetailField.vue";
+import FormField from "./components/FormField.vue";
+
+Nova.booting((Vue, router, store) => {
+    Nova.inertia('nova-permissions', Tool)
+
+    Vue.component('index-permission-checkboxes', IndexField)
+    Vue.component('detail-permission-checkboxes', DetailField)
+    Vue.component('form-permission-checkboxes', FormField)
 })