+++ src/Http/Controllers/ToolController.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace :namespace_vendor\:namespace_tool_name\Http\Controllers;
+namespace Chilly2go\NovaPermissions\Http\Controllers;
 
 use Illuminate\Http\Request;
-use :namespace_vendor\:namespace_tool_name\File;
+use Chilly2go\NovaPermissions\File;
 use Illuminate\Routing\Controller;
 
 class ToolController extends Controller